Improve high-quality icons
Now high-quality icons are included not by using a macro, but in a
configuration file. To do this add the `use_hq_icons` key to the
`inventory` section of `system.ltx` file. If value is `true`, icons
100x100 will be used, otherwise the original 50x50

Also a file was created to read the values from the configuration
for further verification in xrGame area