Fix shotgun animations
These edits are not necessary for the correct functioning of the
game and do not need to be combined with other weapon packs in any way.
Fixed spas12 and winchester1300 animations. Earlier in IX-Ray was
reduced the distance of clipping hud from the first person
and because of this came out old bugs, in particular was visible shoulder