Fix shotgun animations
These edits are not necessary for the correct functioning of the
game and do not need to be combined with other weapon packs in any way.
Fixed spas12 and winchester1300 animations.
Earlier in IX-Ray the distance of first person hud clipping was reduced
and because of that bugs like visible shoulder were occurring