Implement automatic invalid `level_vertex_id` correction in paths
The correction is made when using this path if it refers to the
current location. I left path() unchanged and added safe_path(),
in which this correction is made. Where necessary, I changed path()
to safe_path(). I did not change it in alifetask, as only game_vertex_id
is used there. Determining whether a given path refers to the current
location is done by its game_vertex_id. As soon as the path has a
point with an invalid game_vertex_id or does not refer to the current
location, it means that this path is not in the current location and does not need to be corrected
. In safe_path(), I added a third parameter that shows
that this path is definitely in the current location and we can do without
checking game_vertex_id, which can be invalid because it is simply not
used in some cases. For example, it is not used for look paths.
You can tell that the path is definitely in this location by You can tell that the path is definitely at this location by
where it is used. For example, for home mutants, the path is definitely at the
current location, as it only applies to online mutants

Co-Authored-By: Denis Shaposhnikov <993498+dsh2dsh@users.noreply.github.com>